+++ changelog.md
@@ -4,3 +4,4 @@
 - Restructure files to follow pypi structure
 - Make BloodHound 4.1 patch optional
 - Add changelog
+- Add more options relating to logging and output